Split code into sub-modules
